requirements.txt
streamlit==1.52.2
tensorflow==2.20.0
opencv-python==4.12.0.88
numpy==2.2.6
Pillow==12.1.0
scikit-learn==1.8.0
tqdm==4.67.1
kagglehub==0.3.13